Move all Postgres scripts into postgres adapter package
