+++ lib/resource-definition/generate.js
@@ -1,9 +1,9 @@
 'use strict';
 
 const _ = require('lodash');
-const sqlUtil = require('../sql/sql-util');
 const buildJsonSchema = require('./build-json-schema');
 const relationshipUtil = require('../relationship-util');
+const sqlUtil = require('../../packages/api-pls-postgres-adapter/sql/sql-util');
 
 // Some of the properties of the Resource Models are objects, because that
 // minimizes the typing that a human has to do. But they are more useful to us
