+++ packages/api-pls-postgres-adapter/middleware/create.js
@@ -3,7 +3,7 @@
 const _ = require('lodash');
 const handleQueryError = require('../util/handle-query-error');
 const formatTransaction = require('../util/format-transaction');
-const crud = require('../../../lib/sql/crud');
+const crud = require('../sql/crud');
 const serverErrors = require('../../../lib/server-errors');
 
 module.exports = async function(req) {
