+++ packages/api-pls-postgres-adapter/sync/util/apply.js
