+++ packages/api-pls-postgres-adapter/util/build-response-relationships.js
@@ -1,8 +1,8 @@
 'use strict';
 
 const _ = require('lodash');
-const sqlUtil = require('../../../lib/sql/sql-util');
-const manyToManyUtil = require('../../../lib/sql/many-to-many-util');
+const sqlUtil = require('../sql/sql-util');
+const manyToManyUtil = require('../sql/many-to-many-util');
 const relationshipUtil = require('../../../lib/relationship-util');
 
 function formatToOneResult({result, definition, value, version, columnBase, relation, adjustResourceQuantity}) {
